// src/api/endpoints/api.schemas.ts
export interface RoomInRenovation {
  id?: string;
  roomId?: string;
  renovationId?: string;
  x?: number;
  y?: number;
  z?: number;
  rotationX?: number;
  rotationY?: number;
  rotationZ?: number;
  scaleX?: number;
  scaleY?: number;
  scaleZ?: number;
  visible?: boolean;
}